feat: implement ingestion orchestration and progress tracking
- Added orchestrator module for handling the ingestion process, including functions for executing ingestion with various configurations.
- Introduced progress tracking with emit_progress and emit_checkpoint_ingested functions to report the status of the ingestion process.
- Implemented semantic refresh functionalities, including selecting active code embedding states and running semantic feature refresh.
- Created shared utilities for resolving pack versions, tracking paths at revisions, and promoting temporary rows for head commits.
